refactor(runner): split agent run tests by contract
