+++ crates/runner/src/executor/tests/agent_run_tests/mod.rs
@@ -1,0 +1,8 @@
+mod active_input;
+mod bootstrap;
+mod cancellation;
+mod codex_model_catalog_prefetch;
+mod guest_setup;
+mod session_history;
+mod storage;
+mod support;
